% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aSPU.R
\name{correspLeaves}
\alias{correspLeaves}
\title{Finding the correspondence OTUs for each taxa }
\usage{
correspLeaves(tree)
}
\arguments{
\item{tree}{Rooted phylogenetic tree of R class \dQuote{phylo}}
}
\value{
A list containing the descendants for each taxa.
}
\description{
	Finding the descendants for each taxa. 
}
\examples{
data(throat.tree)
correspLeaves(throat.tree)
}
\author{
Chong Wu
}
\references{
Chong, W., Pan, W. (2015) An Adaptive Association Test for Microbiome Data, submitted. 
}


