% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aSPU.R
\name{simulateData}
\alias{simulateData}
\title{OTU counts simulation}
\usage{
simulateData(nSam=100, s=12, ncluster = 20, mu = 1000, size = 25) 
}
\arguments{
\item{nSam}{Sample size, the default value is 100.}
\item{s}{The indicator of informative cluster.}
\item{ncluster}{Positive integer specifying the number of clusters of they phylogenetic tree.}
\item{mu}{The mean of the negative binomial distribution. }
\item{size}{The size of the negative binomial distribution}
}
\value{
A list object, including the informative OTU counts table and whole OTU counts table.
}
\description{
We used a phylogenetic tree of OTUs from a real throat microbiome data set, which consists of 856 OTUs after discarding singleton OTUs. Then based on a complicate statistical meodel, we generated the OTU counts for each individual to simulate the feature of real microbiome data.
}
\examples{
OTU = simulateData()
OTU
}
\author{
Chong Wu
}
\references{
Chong, W., Pan, W. (2015) An Adaptive Association Test for Microbiome Data, submitted. 
}


